Fixed the error introduced from last commit, where enabling dragging from library to drop to playlist broke the drag and drop within playlist itself.
The problem seems to be the DnD system can detect only one type of
mimedata for dropping. Fixed this by making both items from library and
playlist have mimeData="myMediaItem", then checking the header inserted
into the QByteArray, which is either the QString "libraryItem" or
"playlistItem".

The header is then used in dropEvent in playlistTable to trigger
different drop actions.