Library and Database persistence now works fully:
All paths that have been used to do "Import from Folder" are remembered
in a text file "AAMusicPlayer_CONFIG.txt".

On launching the music player, it will first populate the library
according to the existing database items in "AAMusicPlayer_library.db3",
deleting entries if they no longer exist.

Then it will go through the directories listed in
"AAMusicPlayer_CONFIG.txt" and add into the database and library any
files not already present.

If during the operation of the music player, any files are removed or
added to any of the directories in ""AAMusicPlayer_CONFIG.txt" (in
addition to other directories used to imported from during that
session), clicking the "File->Refresh Library" button will refresh the
library and show the updates.